fix: repository project structure

+++ src/eu/derzauberer/pis/service/EntityService.java
@@ -19,18 +19,30 @@ public abstract class EntityService<T extends Entity<T>> {
 	}
 	
 	public void add(T entity) {
+		if (entity.getId() == null || entity.getId().isEmpty()) {
+			throw new IllegalArgumentException("Entity id must be not null and not empty!");
+		}
 		repository.add(entity);
 	}
 	
 	public boolean removeById(String id) {
+		if (id == null || id.isEmpty()) {
+			throw new IllegalArgumentException("Id must be not null and not empty!");
+		}
 		return repository.removeById(id);
 	}
 	
 	public boolean containsById(String id) {
+		if (id == null || id.isEmpty()) {
+			throw new IllegalArgumentException("Id must be not null and not empty!");
+		}
 		return repository.containsById(id);
 	}
 	
 	public Optional<T> getById(String id) {
+		if (id == null || id.isEmpty()) {
+			throw new IllegalArgumentException("Id must be not null and not empty!");
+		}
 		return repository.getById(id);
 	}
